Extract to new class project descriptor generation
