+++ spec/services/project_descriptor_factory_spec.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+describe ProjectDescriptorFactory do
+  include_context "basic_context"
+
+  let(:expected_descriptor) { JSON.parse(fixture_content(:scorpio, "project_descriptor.json")) }
+
+  it "should instantiate a valid project" do
+    project = full_project
+
+    project.payment_rules.first.datasets.create!(frequency: "quarterly", external_reference: "aze123")
+
+    descriptor = ProjectDescriptorFactory.new.project_descriptor(project)
+    actual_descriptor = as_json(descriptor)
+    puts JSON.generate(descriptor) if actual_descriptor != expected_descriptor
+
+    expect(actual_descriptor).to eq(expected_descriptor)
+  end
+
+  def as_json(descriptor)
+    JSON.parse(JSON.generate(descriptor))
+  end
+end
